refactor: 컴포넌트 생명주기 관련 패키지 분리
