+++ di/src/main/java/woowacourse/shopping/di/container/ApplicationDependencyContainer.kt
@@ -4,9 +4,9 @@ import android.content.Context
 import androidx.lifecycle.LifecycleOwner
 import woowacourse.shopping.di.ClassQualifier
 import woowacourse.shopping.di.ComponentDependencyProvider
-import woowacourse.shopping.di.ComponentLifecycleObserver
+import woowacourse.shopping.di.lifecycle.ComponentLifecycleObserver
 import woowacourse.shopping.di.DependencyProvider
-import woowacourse.shopping.di.ViewModelLifecycleObserver
+import woowacourse.shopping.di.lifecycle.ViewModelLifecycleObserver
 import woowacourse.shopping.di.annotation.ActivityLifecycleAware
 import woowacourse.shopping.di.annotation.ApplicationLifecycleAware
 import woowacourse.shopping.di.annotation.FragmentLifecycleAware
